Refactor databases protocols
