+++ pwchem/viewers/__init__.py
@@ -1,6 +1,3 @@
-from .viewer_protocol_dali import ProtChemDaliViewer
-from .viewer_protocol_ZL_predict import ProtChemZLPredictViewer
-from .viewer_protocol_raptorX import ProtChemRaptorXViewer
 from .viewers_data import *
 from .viewer_pockets import *
 from .viewer_smallMols import SmallMoleculesViewer
